Fiks en haug med små UU-feil
