+++ web/app/features/article/LikeContent.tsx
@@ -36,7 +36,7 @@ export const LikeContent = ({ id }: LikeContent) => {
           onClick={() => {
             trackEvent('like_content_clicked')
           }}
-          className="px-4 py-2 bg-green-600 hover:bg-green-700 text-white rounded-lg shadow-md transition-colors duration-200 flex items-center gap-2 w-fit mx-auto"
+          className="px-4 py-2 bg-green-700 hover:bg-green-800 text-white rounded-lg shadow-md transition-colors duration-200 flex items-center gap-2 w-fit mx-auto"
         >
           <span className="text-xl">
             <TrendingUp />
