+++ web/app/features/header/Header.tsx
@@ -16,7 +16,7 @@ export const Header = () => {
       className={`grid grid-cols-1 md:grid-cols-[8fr_1fr] p-4 md:pb-14 ${isOnArticlePage ? 'pt-6 px-6 md:pl-10 xl:pl-20 md:pt-5' : 'px-4 md:px-10 md:pt-8'}`}
     >
       <div className={`justify-self-end md:col-start-2 ${error ? 'ml-auto' : ''}`}>
-        <Link to="/post/2024">
+        <Link to="/" aria-label="Gå til forsiden">
           {isOnArticlePage ? (
             <div>
               <PostStamp />
